+++ ChangeLog.md
@@ -10,6 +10,7 @@
 - BUGFIX: #5305 Fixed result counting and syntax assistant in case of manually - by a user - attached databases (using explicit ATTACH statement).
 - BUGFIX: #5311 Trigger list and Index list in Table Window are now sortable and initially sorted by name. Editing them does not change their order (unless it reflects current sortig order).
 - BUGFIX: #5333 Fixed handling dollar sign in object names.
+- BUGFIX: #5314 Fixed proper handling od the USING clause in queries with JOINs.
 - BUGFIX: Fixed arbitrary crashes when opening SQL Editor, while having non-thread-safe SQLite extensions loaded in the active database.
 
 ### 3.4.17